Add FORCE_SSL variable to force SSL behind AWS ELB
If the new FORCE_SSL variable is True, the HAProxy config will
check if the X-Forwarded-Proto header is 'http'. If it is, it
will send a 301 redirect to https.

This is useful when HAProxy is run as the reverse proxy behind
an AWS Elastic Load Balancer, as one can terminate SSL at the ELB
while still forcing HTTPS for all requests.